waiting for last attack to execute before performing the next
it keeps only the last input pressed (as tested in Bayonetta 2)
meaning, if you are in x and press xy it only keeps the last y. you need to press the second x in the animation of the first x and then the y in the animation of the second x to perform the xxy combo.